Move code completions handler into completions codebase
