+++ vscode/src/completions/client.test.ts
